tweak output classes

+++ src/output/album.go
@@ -4,6 +4,6 @@ type Album struct {
 	Artist string  `json:"artist"`
 	Album  string  `json:"album"`
 	MBID   string  `json:"mbid"`
-	Score  float32 `json:"score"`
+	Score  float64 `json:"score"`
 	Tracks []Track `json:"tracks"`
 }
